requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
opencv-python>=4.5.0
pillow>=8.0.0
pyyaml>=6.0
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0

# YOLOv8 and related
ultralytics>=8.0.0

# Data augmentation
albumentations>=1.1.0

# Hyperparameter optimization
optuna>=2.10.0

# API interaction and utilities
python-dotenv>=0.19.0
requests>=2.25.0
aiohttp>=3.8.0
asyncio>=3.4.3

# For ensemble methods
scipy>=1.7.0

# Development tools
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0

# Detection models
effdet==0.4.1  # For EfficientDet

# Geospatial processing
geopy>=2.2.0
pyproj>=3.0.0
geopandas>=0.9.0
shapely>=1.7.0
polyline>=1.4.0
contextily==1.3.0

# Visualization
folium>=0.12.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components==1.4.1

# Additional dependencies
googlemaps>=3.1.0
optuna>=3.0.0

